Moved client related files in to client directory.
Makes room for the service classes.
