+++ spec/brreg_grunndata/client/sok_enhet_query_to_xml_spec.rb
@@ -1,0 +1,34 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+module BrregGrunndata
+  class Client
+    describe SokEnhetQueryToXml do
+      subject do
+        described_class.new 'STATOIL',
+                            first: 10
+      end
+
+      describe '#cdata' do
+        it 'returns expected XML string with cdata' do
+          expected = <<~XML
+            <![CDATA[<BrAixXmlRequest RequestName="BrErfrSok">
+            <BrErfrSok>
+            <BrSokeStreng>STATOIL</BrSokeStreng>
+            <MaxTreffReturneres>10</MaxTreffReturneres>
+            <ReturnerIngenHvisMax>true</ReturnerIngenHvisMax>
+            <RequestingIPAddr>010.001.052.011</RequestingIPAddr>
+            <RequestingTjeneste>SOAP</RequestingTjeneste>
+            <MedUnderenheter>true</MedUnderenheter>
+            </BrErfrSok>
+            </BrAixXmlRequest>
+            ]]>
+          XML
+
+          expect(subject.cdata).to eq expected.delete "\n"
+        end
+      end
+    end
+  end
+end
